refactor: delete barrel files and merge lib into utils
Remove all 19 one-line re-export index.ts barrels, flatten single-file
component directories so former barrel paths resolve to concrete files
(ui/button/button.tsx -> ui/button.tsx), hoist svg icons beside curve,
and move cn from lib/utils into the single utils tree. Every helper now
has exactly one import path. Interface declarations in touched files
converted to type aliases per repo TS conventions.

Closes #34
