+++ src/blocks/CallToAction/call-to-action.tsx
@@ -2,10 +2,10 @@ import Image from 'next/image'
 import Link from 'next/link'
 import type { FC } from 'react'
 import { RichText } from '@/components/rich-text'
-import { Button } from '@/components/ui'
-import { cn } from '@/lib/utils'
+import { Button } from '@/components/ui/button'
+import { cn } from '@/utils/cn'
 import type { CallToActionBlock } from '@/types/payload-types'
-import { getImage } from '@/utils'
+import { getImage } from '@/utils/getImage'
 import { getLink } from '@/utils/getLink'
 import { mapBackgroundColor } from '@/utils/mapBackgroundColor'
