+++ src/blocks/ImageWithText/section-heading.tsx
@@ -1,5 +1,5 @@
 import type { FC } from 'react'
-import { cn } from '@/lib'
+import { cn } from '@/utils/cn'
 import type { ImageWithTextBlock } from '@/types/payload-types'
 
 type Props = {
