--- src/components/accordion/index.ts
@@ -1,1 +1,0 @@
-export * from './accordion'
