--- src/components/disclaimer/index.tsx
@@ -1,1 +1,0 @@
-export * from './disclaimer'
