--- src/components/google-analytics/index.ts
@@ -1,1 +1,0 @@
-export { GoogleAnalyticsComponent } from './google-analytics'
