--- src/components/navigation/index.ts
@@ -1,2 +1,0 @@
-export * from './main-nav'
-export * from './mobile-nav'
