--- src/components/privacy-policy/index.ts
@@ -1,1 +1,0 @@
-export * from './privacy-policy'
