--- src/components/rich-text/index.ts
@@ -1,1 +1,0 @@
-export * from './rich-text'
