--- src/components/site-footer/index.ts
@@ -1,1 +1,0 @@
-export * from './site-footer'
