--- src/components/site-header/index.ts
@@ -1,1 +1,0 @@
-export * from './site-header'
