--- src/components/terms-of-service/index.ts
@@ -1,1 +1,0 @@
-export * from './terms-of-service'
