--- src/components/ui/button/index.ts
@@ -1,1 +1,0 @@
-export * from './button'
