--- src/components/ui/carousel/index.ts
@@ -1,1 +1,0 @@
-export * from './carousel'
