--- src/components/ui/divider/index.ts
@@ -1,1 +1,0 @@
-export * from './divider'
