--- src/components/ui/index.ts
@@ -1,9 +1,0 @@
-export * from './accordion'
-export * from './button'
-export * from './card'
-export * from './carousel'
-export * from './dialog'
-export * from './divider'
-export * from './sheet'
-export * from './svg'
-export * from './word-carousel'
