--- src/components/ui/sheet/index.ts
@@ -1,1 +1,0 @@
-export * from './sheet'
