--- src/components/ui/svg/icons/index.ts
@@ -1,5 +1,0 @@
-export * from './arrow-left'
-export * from './arrow-right'
-export * from './chevron-down'
-export * from './hamburger-menu'
-export * from './x'
