--- src/components/ui/word-carousel/index.ts
@@ -1,1 +1,0 @@
-export * from './word-carousel'
