--- src/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './canUseDOM'
-export * from './getImage'
-export * from './getURL'
